requirements.txt
# Policy Compliance Guardian - Requirements

# Core Dependencies
# python>=3.10

# Google Cloud & AI APIs
google-genai
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-cloud-storage
google-adk

# Web Scraping & HTTP
requests
beautifulsoup4
lxml

# Email & Communication
google-cloud-logging

# Data Processing
python-dotenv
pydantic

# Utilities
loguru
pytz

# Development & Testing
pytest
pytest-cov
pytest-asyncio
black
flake8
mypy
# Optional: For production deployment
gunicorn
flask

# Optional: For OpenTelemetry observability (mentioned in architecture)
# opentelemetry-api
# opentelemetry-sdk
# google-cloud-trace

flask
werkzeug